+++ packages/ng/demo/app/user/tile/displayFormat/display-format.component.ts
@@ -11,7 +11,6 @@ export class DisplayFormatComponent implements OnInit {
 	anais: IUser = {
 		firstName: 'Anaïs',
 		lastName: 'Lemoustier',
-		picture: { href: 'https://upload.wikimedia.org/wikipedia/commons/e/ec/Ana%C3%AFs_Demoustier_Cabourg_2015.jpg' },
 		jobTitle: 'Actress'
 	};