+++ packages/ng/src/app/user/picture/user-picture.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, HostBinding } from '@angular/core';
 import { IUser } from '../user.model';
-import { LuUserDisplayPipe } from '../display';
+import { LuUserDisplayPipe, DisplayFormat, DisplayInitials, DisplayFullname } from '../display';
 
 /**
  * Displays user'picture or a placeholder with his/her initials and random bg color'
@@ -11,6 +11,11 @@ import { LuUserDisplayPipe } from '../display';
 	styleUrls: ['./user-picture.component.scss'],
 })
 export class LuUserPictureComponent {
+	/**
+		 * User Display format.
+		 * It is set to 'FL' by default
+		 */
+	@Input() displayFormat: DisplayInitials = DisplayInitials.firstlast;
 
 	/**
 	 * IUser whose picture you wanna display.
@@ -19,7 +24,7 @@ export class LuUserPictureComponent {
 
 	@Input() set user(user: IUser) {
 		this._user = user;
-		this.initials = this.displayPipe.transform(user, 'LF');
+		this.initials = this.displayPipe.transform(user, this.displayFormat);
 		this.hasPicture = !!user.picture && !!user.picture.href;
 		if (this.hasPicture) {
 			this.style = { 'background-image': `url('${this._user.picture.href}')` };
@@ -29,6 +34,7 @@ export class LuUserPictureComponent {
 		}
 	}
 	get user() { return this._user; }
+
 	initials = '';
 	hasPicture = false;
 
@@ -38,7 +44,7 @@ export class LuUserPictureComponent {
 
 	private getNameHue() {
 		// we sum the chars in user's firstname + lastname
-		const charSum = this.displayPipe.transform(this._user, 'lf')
+		const charSum = this.displayPipe.transform(this._user, DisplayFullname.firstlast)
 		.split('')
 		.reduce((sum, a) => sum + a.charCodeAt(0), 0);
 		// and take a modulo 360 for hue