+++ packages/ng/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## In Dev
 ### New features
+- `lu-user-picture` now has a displayFormat attribute. Default to "FL"
 - `user-select`: Add component for user select
 - `user-select-picker`: Add a picker that implements IOptionFeeder for the user select
 - `IOptionFeeder`: Add a method `textValue` to let developer set how to display the value at first (when no options are loads)